+++ internal/model/store_bargain.gen.go
@@ -30,15 +30,15 @@ type StoreBargain struct {
 	BargainMaxPrice decimal.Decimal       `gorm:"column:bargain_max_price;type:decimal(8,2) unsigned;comment:用户每次砍价的最大金额" json:"bargain_max_price"`                // 用户每次砍价的最大金额
 	BargainMinPrice decimal.Decimal       `gorm:"column:bargain_min_price;type:decimal(8,2) unsigned;comment:用户每次砍价的最小金额" json:"bargain_min_price"`                // 用户每次砍价的最小金额
 	BargainNum      int64                 `gorm:"column:bargain_num;type:int unsigned;not null;default:1;comment:帮砍次数——单个商品用户可以帮砍的次数" json:"bargain_num"`          // 帮砍次数——单个商品用户可以帮砍的次数
-	Status          int64                 `gorm:"column:status;type:tinyint unsigned;not null;default:1;comment:砍价状态 0(到砍价时间不自动开启)  1(到砍价时间自动开启时间)" json:"status"` // 砍价状态 0(到砍价时间不自动开启)  1(到砍价时间自动开启时间)
+	Status          int32                 `gorm:"column:status;type:tinyint unsigned;not null;default:1;comment:砍价状态 0(到砍价时间不自动开启)  1(到砍价时间自动开启时间)" json:"status"` // 砍价状态 0(到砍价时间不自动开启)  1(到砍价时间自动开启时间)
 	GiveIntegral    int64                 `gorm:"column:give_integral;type:int;comment:反多少积分" json:"give_integral"`                                                // 反多少积分
 	Info            string                `gorm:"column:info;type:varchar(255);comment:砍价活动简介" json:"info"`                                                        // 砍价活动简介
 	Cost            decimal.Decimal       `gorm:"column:cost;type:decimal(8,2) unsigned;comment:成本价" json:"cost"`                                                  // 成本价
 	Sort            int64                 `gorm:"column:sort;type:int unsigned;not null;comment:排序" json:"sort"`                                                   // 排序
-	IsHot           int64                 `gorm:"column:is_hot;type:tinyint unsigned;not null;comment:是否推荐0不推荐1推荐" json:"is_hot"`                                  // 是否推荐0不推荐1推荐
-	IsDel           int64                 `gorm:"column:is_del;type:tinyint unsigned;not null;comment:是否删除 0未删除 1删除" json:"is_del"`                                // 是否删除 0未删除 1删除
+	IsHot           int32                 `gorm:"column:is_hot;type:tinyint unsigned;not null;comment:是否推荐0不推荐1推荐" json:"is_hot"`                                  // 是否推荐0不推荐1推荐
+	IsDel           int32                 `gorm:"column:is_del;type:tinyint unsigned;not null;comment:是否删除 0未删除 1删除" json:"is_del"`                                // 是否删除 0未删除 1删除
 	AddTime         int64                 `gorm:"column:add_time;type:bigint unsigned;comment:添加时间" json:"add_time"`                                               // 添加时间
-	IsPostage       int64                 `gorm:"column:is_postage;type:tinyint unsigned;not null;default:1;comment:是否包邮 0不包邮 1包邮" json:"is_postage"`              // 是否包邮 0不包邮 1包邮
+	IsPostage       int32                 `gorm:"column:is_postage;type:tinyint unsigned;not null;default:1;comment:是否包邮 0不包邮 1包邮" json:"is_postage"`              // 是否包邮 0不包邮 1包邮
 	Postage         decimal.Decimal       `gorm:"column:postage;type:decimal(10,2) unsigned;comment:邮费" json:"postage"`                                            // 邮费
 	Rule            string                `gorm:"column:rule;type:text;comment:砍价规则" json:"rule"`                                                                  // 砍价规则
 	Look            int64                 `gorm:"column:look;type:int unsigned;comment:砍价商品浏览量" json:"look"`                                                       // 砍价商品浏览量
