+++ internal/model/system_admin.gen.go
@@ -17,10 +17,10 @@ type SystemAdmin struct {
 	Roles      string                `gorm:"column:roles;type:varchar(128);not null;comment:后台管理员权限(menus_id)" json:"roles"`              // 后台管理员权限(menus_id)
 	LastIP     string                `gorm:"column:last_ip;type:varchar(16);comment:后台管理员最后一次登录ip" json:"last_ip"`                        // 后台管理员最后一次登录ip
 	LoginCount int64                 `gorm:"column:login_count;type:int unsigned;not null;comment:登录次数" json:"login_count"`               // 登录次数
-	Level      int64                 `gorm:"column:level;type:tinyint unsigned;not null;default:1;comment:后台管理员级别" json:"level"`          // 后台管理员级别
-	Status     int64                 `gorm:"column:status;type:tinyint unsigned;not null;default:1;comment:后台管理员状态 1有效0无效" json:"status"` // 后台管理员状态 1有效0无效
+	Level      int32                 `gorm:"column:level;type:tinyint unsigned;not null;default:1;comment:后台管理员级别" json:"level"`          // 后台管理员级别
+	Status     int32                 `gorm:"column:status;type:tinyint unsigned;not null;default:1;comment:后台管理员状态 1有效0无效" json:"status"` // 后台管理员状态 1有效0无效
 	Phone      string                `gorm:"column:phone;type:varchar(15);comment:手机号码" json:"phone"`                                     // 手机号码
-	IsSms      int64                 `gorm:"column:is_sms;type:tinyint unsigned;comment:是否接收短信" json:"is_sms"`                            // 是否接收短信
+	IsSms      int32                 `gorm:"column:is_sms;type:tinyint unsigned;comment:是否接收短信" json:"is_sms"`                            // 是否接收短信
 	CreatedAt  int64                 `gorm:"column:created_at;type:bigint" json:"created_at"`
 	UpdatedAt  int64                 `gorm:"column:updated_at;type:bigint" json:"updated_at"`
 	DeletedAt  soft_delete.DeletedAt `gorm:"column:deleted_at;type:bigint" json:"-"`
