+++ internal/model/user_address.gen.go
@@ -22,8 +22,8 @@ type UserAddress struct {
 	PostCode  int64                 `gorm:"column:post_code;type:int;not null;comment:邮编" json:"post_code"`                           // 邮编
 	Longitude string                `gorm:"column:longitude;type:varchar(16);not null;default:0;comment:经度" json:"longitude"`         // 经度
 	Latitude  string                `gorm:"column:latitude;type:varchar(16);not null;default:0;comment:纬度" json:"latitude"`           // 纬度
-	IsDefault int64                 `gorm:"column:is_default;type:tinyint unsigned;not null;comment:是否默认" json:"is_default"`          // 是否默认
-	IsDel     int64                 `gorm:"column:is_del;type:tinyint unsigned;not null;comment:是否删除" json:"is_del"`                  // 是否删除
+	IsDefault int32                 `gorm:"column:is_default;type:tinyint unsigned;not null;comment:是否默认" json:"is_default"`          // 是否默认
+	IsDel     int32                 `gorm:"column:is_del;type:tinyint unsigned;not null;comment:是否删除" json:"is_del"`                  // 是否删除
 	CreatedAt int64                 `gorm:"column:created_at;type:bigint" json:"created_at"`
 	UpdatedAt int64                 `gorm:"column:updated_at;type:bigint" json:"updated_at"`
 	DeletedAt soft_delete.DeletedAt `gorm:"column:deleted_at;type:bigint" json:"-"`
